+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Binary/Golden.hs
@@ -103,9 +103,9 @@ goldenNewEpochStateExpectation
             , Ev ver esLState
             , Em
                 [ E (TkListLen 4)
-                , snapShotEnc ssStakeMark
-                , snapShotEnc ssStakeSet
-                , snapShotEnc ssStakeGo
+                , markSnapShotEnc ssStakeMark
+                , setSnapShotEnc ssStakeSet
+                , goSnapShotEnc ssStakeGo
                 , E ssFee
                 ]
             , Ev ver esNonMyopic
@@ -128,6 +128,27 @@ goldenNewEpochStateExpectation
           [ E (TkListLen 2)
           , Ev ver ssActiveStake
           , Ev ver ssStakePoolsSnapShot
+          ]
+      markSnapShotEnc MarkSnapShot {..} =
+        Em
+          [ E (TkListLen 3)
+          , snapShotEnc msSnapShot
+          , Ev ver msEpochNo
+          , Ev ver msLeiosCommitteeSize
+          ]
+      -- ssPoolDistr and ssLeiosCommittee are derived, so not serialized.
+      setSnapShotEnc SetSnapShot {..} =
+        Em
+          [ E (TkListLen 3)
+          , snapShotEnc ssSnapShot
+          , Ev ver ssEpochNo
+          , Ev ver ssLeiosCommitteeSize
+          ]
+      -- gsPoolDistr is derived, so not serialized.
+      goSnapShotEnc GoSnapShot {..} =
+        Em
+          [ E (TkListLen 1)
+          , snapShotEnc gsSnapShot
           ]
 
 shelleyDecodeDuplicateDelegCertSucceeds :: Version -> Spec